feat(frontend): restore contact edit, merge and archive on the person record page
margince#4823: PersonPageV2 — the route every /#/contacts/{id} open actually
resolves to since PersonScreen's own routing moved there — offered no way
to correct a bounced email, a wrong phone number, a misspelled name or a
title. The working edit/merge/archive form still existed, fully tested,
hanging off PersonScreen (contacts.tsx), which nothing routes to any more.
API-only was the only path left for what should be an ordinary correction.

PersonEditMergeArchive is the shared component both pages now render into
their header's badges slot: one writer of the PATCH/merge/archive wiring
rather than two, wired to whichever caller's own write-refusal reason is
correct for it — PersonScreen's simpler archived-only check, PersonPageV2's
useRecordWriteRefusal (which also answers "not yours to change", a case
the old page's own header verbs never had to consider).

Splitting this out — and PersonActions, personformfields.ts's create/edit
field schema and request mappers, and the shared personpage.testkit.tsx
fixture — is what let contacts.tsx and personpage.tsx absorb the wiring
without growing past the frontend's file-length ratchet: each of the three
touched files is smaller after this change than the waiver that governed
it before, not merely no larger.
